Extracted compiler
